refs #438 define -DASIO_DISABLE_TIMERFD to explicitly disable timerfd usage in asio epoll reactor (+ build script fix)
git-svn-id: https://devel.codership.com/svn/galera/branches/0.8@2073 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ scripts/mysql/build.sh
@@ -337,7 +337,10 @@ then
 # this.
 install -m 755 -D $MYSQL_SRC/libmysql/.libs/libmysqlclient.so $MYSQL_LIBS/libmysqlclient.so.16
 fi
+if test -f $MYSQL_SRC/storage/innodb_plugin/.libs/ha_innodb_plugin.so
+then
 install -m 755 -D $MYSQL_SRC/storage/innodb_plugin/.libs/ha_innodb_plugin.so $MYSQL_PLUGINS/ha_innodb_plugin.so
+fi
 install -m 755 -d $MYSQL_BINS
 if [ "$SKIP_CLIENTS" == "no" ]
 then
